Make Buy/Sell parse on demand
The filters concept was doing really bad because the original menu baked stuff in, and that list was cloned into the menu's `options` list.  Both needed to be altered, and the component render updates were non-obvious, and still ultimately relied on reactivity to `inventory` to come back around after the events fired.

This way, we just treat `items` as an optional callable, and use it to build a fresh list anytime it's needed, which circumvents racing the inventory events during useEffect.